+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 * Rake task for migration to be implemented in 'rom-migrator' (nepalez)
 
+### Internal
+
+* Gateway isolates session from dataset (nepalez)
+
 [Compare v0.0.2...v0.1.0](https://github.com/rom-rb/rom-cassandra/compare/v0.0.2...v0.1.0)
 
 ## v0.0.2 2015-09-08